+++ functions/src/services/page-context.service.ts
@@ -1,8 +1,8 @@
 import { QuerySnapshot } from '@google-cloud/firestore';
 import * as admin from 'firebase-admin';
 import { Document, DocumentContext, PageContext, SiteContext, SiteInformation } from '../models';
-import * as documentService from '../services/document.service';
-import * as systemNotificationService from '../services/system-message.service';
+import * as documentService from './document.service';
+import * as systemNotificationService from './system-message.service';
 export interface DocumentQueryOptions {
     limit?: number;
     orderBy?: {
@@ -14,13 +14,13 @@ export interface DocumentQueryOptions {
 const siteCollection = () => admin.firestore().collection('tanam').doc(process.env.GCLOUD_PROJECT);
 
 
-export async function queryDocumentContext(documentTypeId: string, queryOpts: DocumentQueryOptions = {}) {
-    console.log(`[queryDocumentContext] ${documentTypeId}, query=${JSON.stringify(queryOpts)}`);
+export async function queryPageContext(documentTypeId: string, queryOpts: DocumentQueryOptions = {}) {
+    console.log(`[queryPageContext] ${documentTypeId}, query=${JSON.stringify(queryOpts)}`);
     const orderByField = queryOpts.orderBy && queryOpts.orderBy.field || 'updated';
     const sortOrder = queryOpts.orderBy && queryOpts.orderBy.sortOrder || 'desc';
     const limit = queryOpts.limit || 20;
 
-    console.log(`[queryDocumentContext] effective query ${JSON.stringify({ orderByField, sortOrder, limit })}`);
+    console.log(`[queryPageContext] effective query ${JSON.stringify({ orderByField, sortOrder, limit })}`);
     let querySnap: QuerySnapshot;
     try {
         querySnap = await siteCollection()
@@ -32,32 +32,32 @@ export async function queryDocumentContext(documentTypeId: string, queryOpts: Do
             .get();
 
     } catch (err) {
-        console.error(`[queryDocumentContext] ${JSON.stringify(err)}`);
+        console.error(`[queryPageContext] ${JSON.stringify(err)}`);
         await systemNotificationService.reportMissingIndex(err.details);
         return [];
     }
 
-    console.log(`[queryDocumentContext] num results: ${querySnap.docs.length}`);
+    console.log(`[queryPageContext] num results: ${querySnap.docs.length}`);
 
     const result = [];
     for (const doc of querySnap.docs) {
-        console.log(`[queryDocumentContext] ${JSON.stringify(doc.data())}`);
+        console.log(`[queryPageContext] ${JSON.stringify(doc.data())}`);
         result.push(_toContext(doc.data() as Document));
     }
 
     return result;
 }
 
-export async function getDocumentContextById(docId: string) {
-    console.log(`[getDocumentContextById] ${JSON.stringify(docId)}`);
+export async function getPageContextById(docId: string) {
+    console.log(`[getPageContextById] ${JSON.stringify(docId)}`);
     const doc = await documentService.getDocumentById(docId);
     return !!doc ? _toContext(doc) : null;
 }
 
-export async function getDocumentContextByUrl(url: string): Promise<PageContext> {
-    console.log(`[getDocumentContextByUrl] ${JSON.stringify(url)}`);
+export async function getPageContextByUrl(url: string): Promise<PageContext> {
+    console.log(`[getPageContextByUrl] ${JSON.stringify(url)}`);
     const documents = await documentService.getDocumentByUrl(url || '/');
-    console.log(`[getDocumentContextByUrl] Number of query results: ${documents.length}`);
+    console.log(`[getPageContextByUrl] Number of query results: ${documents.length}`);
     return documents.length === 0 ? null : _toContext(documents[0]);
 }
